+++ src/utils/isPageUid.ts
@@ -1,0 +1,4 @@
+export const isPageUid = (uid: string) =>
+  !!window.roamAlphaAPI.pull("[:node/title]", [":block/uid", uid])?.[
+    ":node/title"
+  ];
